Refactored singin.js into signin_functions.js
