+++ app/checkin.html
@@ -42,7 +42,7 @@
   <button id="startFlowBtn" style="display: block;">Start My Flow</button>
   <button id="cancelBtn" style="display: block;">Cancel</button>
 
-  
+
   <script> require('./js/checkin.js') </script>
 
 </body>
