+++ app/js/signin_functions.js
@@ -1,0 +1,9 @@
+function guidVal() {
+    const s4 = () => Math.floor((1 + Math.random()) * 0x10000)
+        .toString(16)
+        .substring(1);
+    // return id of format 'aaaaaaaa'-'aaaa'-'aaaa'-'aaaa'-'aaaaaaaaaaaa'
+    return `${s4() + s4()}-${s4()}-${s4()}`;
+}
+
+module.exports = {guidVal};
